+++ Setup-Your-Mac-via-Dialog.bash
@@ -10,14 +10,15 @@
 #
 # HISTORY
 #
-#   Version 1.15.0, 28-May-2024
+#   Version 1.15.0, 29-May-2024
 #   - Added logging functions
 #   - Modified Microsoft Teams Message `activitySubtitle`
 #   - Activated main "Setup Your Mac" dialog with each `listitem`
-#   - Added swiftDialog `2.5.0` `--verbose` flag to `verbose` debugMode
+#   - Added swiftDialog `2.5.0`'s `--verbose` and `--debug` flags to debugModes
 #   - Failure Message: Increased `sleep` value from `0.3` to `0.7` (thanks, for the report, @arnoldtaw; thanks for the code suggestion, @jcmbowman)
 #   - Miscellaneous formatting and clean-up
-#   - Add Support Team fields (thanks, @HowardGMac!)
+#   - Added Support Team fields (thanks, @HowardGMac!)
+#   - Set `swiftDialogMinimumRequiredVersion` to `2.5.0.4761`
 #
 ####################################################################################################
 
@@ -33,7 +34,7 @@
 # Script Version and Jamf Pro Script Parameters
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-scriptVersion="1.15.0-b7"
+scriptVersion="1.15.0-b8"
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin
 scriptLog="${4:-"/var/log/org.churchofjesuschrist.log"}"                        # Parameter 4: Script Log Location [ /var/log/org.churchofjesuschrist.log ] (i.e., Your organization's default location for client-side logs)
 debugMode="${5:-"verbose"}"                                                     # Parameter 5: Debug Mode [ verbose (default) | true | false ]
@@ -43,7 +44,7 @@ requiredMinimumBuild="${8:-"disabled"}"                                         
 outdatedOsAction="${9:-"/System/Library/CoreServices/Software Update.app"}"     # Parameter 9: Outdated OS Action [ /System/Library/CoreServices/Software Update.app (default) | jamfselfservice://content?entity=policy&id=117&action=view ] (i.e., Jamf Pro Self Service policy ID for operating system ugprades)
 webhookURL="${10:-""}"                                                          # Parameter 10: Microsoft Teams or Slack Webhook URL [ Leave blank to disable (default) | https://microsoftTeams.webhook.com/URL | https://hooks.slack.com/services/URL ] Can be used to send a success or failure message to Microsoft Teams or Slack via Webhook. (Function will automatically detect if Webhook URL is for Slack or Teams; can be modified to include other communication tools that support functionality.)
 presetConfiguration="${11:-""}"                                                 # Parameter 11: Specify a Configuration (i.e., `policyJSON`; NOTE: If set, `promptForConfiguration` will be automatically suppressed and the preselected configuration will be used instead)
-swiftDialogMinimumRequiredVersion="2.5.0.4760"                                  # This will be set and updated as dependancies on newer features change.
+swiftDialogMinimumRequiredVersion="2.5.0.4761"                                  # This will be set and updated as dependancies on newer features change.
 
 
 
@@ -1768,8 +1769,8 @@ esac
 
 dialogBinary="/usr/local/bin/dialog"
 case ${debugMode} in
-    "true"      ) dialogBinary="${dialogBinary}" ;;
-    "verbose"   ) dialogBinary="${dialogBinary} --verbose" ;;
+    "true"      ) dialogBinary="${dialogBinary} --verbose" ;;
+    "verbose"   ) dialogBinary="${dialogBinary} --verbose --debug red" ;;
 esac